pyproject.toml
[tool.poetry]
name = "cloud_run_etl_template"
version = "0.1.0"
description = "The cloud_run_etl_template package"
readme = "README.md"
license = "MIT"
authors = ["Manuel Castillo <m.castillo@altostratus.es>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
cloudops-logging-google = "^0.1.0"
google-cloud-error-reporting = "^1.5.3"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pandas = "^2.0.3"
pyaml-env = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
